Restructured preferences code
